API addition and move parallel-deploy
I moved parallel-deploy to its own standalone repo because
it doesn't really belong with this repo.  It will however
use the protobuf API that this repo defines.  I also added
some new API calls to allow datacenter info or DCIM integration.
As long as the DCIM supports this protobuf API through a shim
the parallel-deploy tool and ticket creation functions will be
able to leverage it for more accurate support tickets and
ceph deployments.