Better raid detection
There's so many possibly configurations with disks that getting
detection to be accurate has proven challenging.  I have made some
improvements to the block-utils library to make better use of the
kernel raid driver info.  I then expose that info to the state
machine so it can make better decisions.